+++ src/app.js
@@ -7,9 +7,9 @@ const catcher = {
         });
     },
 
-    onClicked: function (info, tab) {
-        chrome.storage.sync.get(null, function (config) {
-            if (!config) {
+    onClicked: function (info) {
+        chrome.storage.sync.get(null, config => {
+            if (0 === Object.keys(config).length) {
                 catcher.notification(
                     chrome.i18n.getMessage('missingConfigurationTitle'),
                     chrome.i18n.getMessage('missingConfigurationText')
@@ -24,49 +24,60 @@ const catcher = {
                 return;
             }
 
-            links.forEach(function (ajfsp) {
-                catcher.handleLink(ajfsp, config);
+            links.forEach(ajfsp => {
+                if (config['mode'] === 'indirect') {
+                    catcher.requestImplicit(ajfsp, config);
+                } else {
+                    catcher.requestDirect(ajfsp, config);
+                }
             });
         });
     },
 
-    handleLink: function (ajfsp, config) {
-        let coreUrl = config['host'];
-        coreUrl = coreUrl.endsWith('/') ? coreUrl.slice(0, -1) : coreUrl;
+    requestImplicit: function (ajfsp, config) {
+        let params = {
+            host: config['host'] + ':' + config['port'],
+            cpass: md5(config['password']),
+            ajfsp_link: ajfsp
+        };
 
-        if (config['mode'] === 'indirect') {
-            let phpGUI = config['phpgui'];
-            phpGUI += (phpGUI.endsWith('/') ? '' : '/');
+        let url = new URL(config['phpgui']);
+        url.pathname = 'main/top.php';
 
-            $.post(phpGUI + 'main/index.php', {
-                host: coreUrl + ':' + config['port'], // new phpGUI version
-                ip: coreUrl,
-                port: config['port'],
-                cpass: config['password']
+        window.fetch(url, {
+            method: 'POST',
+            headers: {'Content-Type': 'application/x-www-form-urlencoded'},
+            body: (new URLSearchParams(params)).toString()
+        })
+            .then(result => {
+                result.text().then(data => {
+                    catcher.handleResult(data, ajfsp);
+                });
             })
-                .done(function (result) {
-                    $.get(phpGUI + 'main/top.php', {ajfsp_link: ajfsp})
-                        .done(function (result) {
-                            catcher.handleResult(result, ajfsp);
-                        })
-                        .fail(function (xhr, status, error) {
-                            catcher.notification('phpGUI connection error', status);
-                        });
-                })
-                .fail(function (xhr, status, error) {
-                    catcher.notification('phpGUI connection error', status);
-                });
-        } else {
-            let url = coreUrl + ':' + config['port'] + '/function/processlink';
-            let pass = CryptoJS.MD5(config['password']);
-            $.get(url, {link: ajfsp, password: pass.toString()})
-                .done(function (result) {
-                    catcher.handleResult(result, ajfsp);
-                })
-                .fail(function (xhr, status, error) {
-                    catcher.notification('Core connection error', status);
-                });
-        }
+            .catch(error => {
+                catcher.notification('phpGUI connection error', error.toString());
+            });
+    },
+
+    requestDirect: function (ajfsp, config) {
+        let params = {
+            link: ajfsp,
+            password: md5(config['password'])
+        };
+
+        let url = new URL(config['host']);
+        url.pathname = '/function/processlink';
+        url.port = config['port'];
+        url.search = new URLSearchParams(params).toString();
+
+        window.fetch(url).then(result => {
+            result.text().then(data => {
+                catcher.handleResult(data, ajfsp);
+            });
+        }).catch(error => {
+            console.log(error);
+            catcher.notification('Core connection error', error.toString());
+        });
     },
 
     handleResult: function (result, link) {
@@ -80,13 +91,15 @@ const catcher = {
 
         if ('ok' === result || result.match(/Download:(.*)ok/)) {
             catcher.notification(chrome.i18n.getMessage('linkCatchedSuccessTitle'), link.split('|')[1]);
+        } else if (result.match(/already downloaded/)) {
+            catcher.notification('already downloaded', link.split('|')[1]);
         } else {
             catcher.notification(chrome.i18n.getMessage('linkCatchedErrorTitle'), link.split('|')[1]);
         }
     },
 
     notification: function (title, text) {
-        chrome.notifications.create('appleJuice', {
+        chrome.notifications.create(null, {
             type: 'basic',
             iconUrl: 'icons/apple256.png',
             title: title,